# project-interface/src/main/resources/api-spec.yaml
openapi: 3.0.3
info:
  title: Seoul Map Service API
  description: 서울시 실시간 지도 서비스 및 혼잡도/날씨 정보 API 명세 (REST)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
tags:
  - name: Map
    description: 지도 및 실시간/예측 데이터 조회
  - name: Auth
    description: 인증(회원가입/로그인)
  - name: User
    description: 사용자 관리 및 알람 설정

paths:
  # 1. 실시간 혼잡도 정보 조회
  /map/current:
    get:
      operationId: getMapData
      tags: [Map]
      summary: 지역별 실시간 혼잡도/날씨/대기질 전체 조회
      description: |
        서울시 주요 지역의 실시간 데이터를 리스트 형태로 조회합니다.
      parameters:
        - name: regionId
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: 특정 지역 필터링 (생략 시 전체 조회)
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MapInfo'

  # 2. 혼잡도 예상정보 조회
  /map/forecast:
    get:
      operationId: getMapForecast
      tags: [Map]
      summary: 지역별 혼잡 예상도 전체 조회
      description: 향후 12시간(또는 특정 기간)의 혼잡도 예측 데이터를 조회합니다.
      parameters:
        - name: regionId
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ForecastInfo'

  # 3. 특정 지역 상세정보 조회
  /map/summary/{regionId}:
    get:
      operationId: getRegionSummary
      tags: [Map]
      summary: 특정 지역 상세 정보 단건 조회
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MapInfo'
        '404':
          description: 존재하지 않는 지역 ID

  # 4. 회원가입
  /auth/signup:
    post:
      operationId: signup
      tags: [Auth]
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: 회원가입 성공 (Created)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: 이미 존재하는 이메일 (Conflict)

  # 5. 로그인
  /auth/signin:
    post:
      operationId: signin
      tags: [Auth]
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 이메일 또는 비밀번호 불일치

  # 6. 로그아웃
  /auth/signout:
    post:
      operationId: signout
      tags: [Auth]
      summary: 로그아웃
      security:
        - bearerAuth: []
      parameters:
        - name: Authorization
          in: header
          description: Bearer Token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 로그아웃 성공
        '401':
          description: 인증되지 않은 사용자

  # 7. 회원탈퇴
  /users/me:
    delete:
      operationId: withdrawUser
      tags: [User]
      summary: 회원 탈퇴
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
                  description: 본인 확인용 비밀번호
      responses:
        '204':
          description: 탈퇴 성공 (반환할 데이터 없음)
        '403':
          description: 비밀번호 불일치 (권한 없음)
        '401':
          description: 인증 토큰 만료 또는 없음

  # 8. 알림 설정
  /users/me/alarms/{regionId}:
    post:
      operationId: setAlarms
      tags: [User]
      summary: 지역 알림 구독 (추가)
      security:
        - bearerAuth: []
      parameters:
        - name: regionId
          in: path
          required: true
          description: 알림을 받을 지역 ID
          schema:
            type: integer
            format: int32
      responses:
        '201':
          description: 알림 설정 완료 (Created)
        '404':
          description: 존재하지 않는 지역 ID

    delete:
      operationId: deleteAlarms
      tags: [User]
      summary: 지역 알림 해제
      security:
        - bearerAuth: []
      parameters:
        - name: regionId
          in: path
          required: true
          description: 알림을 해제할 지역 ID
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: 해제 성공 (No Content)


components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # 1. 통합 맵 정보 DTO
    MapInfo:
      type: object
      properties:
        regionId:
          type: integer
          format: int32
        congestionLevel:
          type: integer
          description: "1:원활, 2:보통, 3:약간혼잡, 4:혼잡"
          enum: [ 1, 2, 3, 4 ]
        weatherCode:
          type: integer
          description: "기상청 코드 기준"
        airQualityIndex:
          type: integer
          description: "1:좋음 ~ 4:나쁨"
          enum: [ 1, 2, 3, 4 ]
        timestamp:
          type: string
          format: date-time

    # 2. 예상 혼잡도 정보 DTO
    ForecastInfo:
      type: object
      properties:
        regionId:
          type: integer
          format: int32
          description: 지역 ID
        congestionLevel:
          type: integer
          format: int32
          description: 혼잡도 예상
          enum: [ 1, 2, 3, 4 ]
        timestamp:
          type: string
          format: date-time
          description: 예측 기준 시간

    # 3. 회원가입 요청 DTO (passwordCheck 제거 권장 반영)
    SignupRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
        username:
          type: string

    # 4. 토큰 응답 DTO
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: "JWT Access Token"
        tokenType:
          type: string
          example: "Bearer"